desktop-app/requirements.txt
customtkinter
esptool
mttkinter
paho_mqtt
pyserial
Requests
